update height type in the `staking` module, use default height type

+++ x/staking/src/keeper/mod.rs
@@ -30,7 +30,7 @@ use gears::{
     x::{keepers::auth::AuthKeeper, module::Module, types::validator::BondStatus},
 };
 use prost::bytes::BufMut;
-use std::{cmp::Ordering, collections::HashMap, u64};
+use std::{cmp::Ordering, collections::HashMap};
 
 // Each module contains methods of keeper with logic related to its name. It can be delegation and
 // validator types.
@@ -555,7 +555,7 @@ impl<
         &self,
         ctx: &mut CTX,
         val_addr: &ValAddress,
-    ) -> Result<(Timestamp, u64, bool), GasStoreErrors> {
+    ) -> Result<(Timestamp, u32, bool), GasStoreErrors> {
         // TODO: When would the validator not be found?
         let validator = self.validator(ctx, val_addr)?;
         let validator_status = validator
@@ -573,7 +573,7 @@ impl<
                 let time =
                     chrono::DateTime::from_timestamp(time.seconds, time.nanos as u32).unwrap();
                 let completion_time = time + duration;
-                let height = ctx.height() as u64;
+                let height = ctx.height();
                 let completion_time = Timestamp {
                     seconds: completion_time.timestamp(),
                     nanos: completion_time.timestamp_subsec_nanos() as i32,
